+++ Database/data.csv
@@ -10,3 +10,5 @@ aah,InMotion-S1,08.05.2020,4512,Pakistan,,3
 aai,Apollo-City,24.11.2021,1452.13,Avrig,geo,1
 aaj,Dual-Storm,08.11.2021,144.3,Turnisor,,0
 aak,Varla-One,05.04.2022,2478.9,Tiglari,,1
+sCn,Saniuta,11.02.1878,300,Neamt,,4
+WRc,AOVO,23.04.2023,10.4,Cluj-Napoca,,0
